pyproject.toml
[project]
name = "agent-web-context"
version = "0.1.0"
description = "MCP server for web browsing, content extraction, and search"
requires-python = ">=3.13"
dependencies = [
    "patchright",
    "beautifulsoup4",
    "fastapi",
    "pydantic",
    "uvicorn[standard]",
    "lxml",
    "aiosqlite",
    "apscheduler",
    "sqlalchemy[asyncio]",
    "aiohttp",
    "mcp",
    "markdownify",
    "langchain",
    "langchain-openai",
    "langchain-anthropic",
    "langchain-ollama",
    "langchain-google-genai",
    "pyyaml>=6.0.2",
    "pymupdf4llm"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
browser = ["patchright"]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.11",
]
